backend/requirements.txt
fastapi
uvicorn
sqlmodel
aiosqlite
python-multipart
pydantic-settings
httpx
pyautogui
Pillow
easyocr
numpy<2
PyMuPDF
python-docx
python-pptx
google-generativeai
psutil
requests
beautifulsoup4
sentence-transformers
faster-whisper
edge-tts
praat-parselmouth
torch
transformers
scipy
sqlalchemy
maturin
pyyaml
